fixed a couple f-string oopsies.
Note about moving back to f-strings for py3.6 in the future...

If you like the PyCharm IDE (I definitely do), you will be glad to hear that it fully supports Python 3.6 and f-strings. Code completion works with expressions embedded inside f-strings just like with ordinary expressions, and there is even a code intention that suggests to “convert to f-string literal”. This means if you have an old Python program that still uses the .format() method, you can just press Alt+Enter and select that intention to convert everything to the new syntax. So you have no more excuses to not switch over.