logging done.
There are two types of logging.

sys_logger reports on main program running details. Normally exists in the main()

logger is a device/thread specific logger for the exectution of procedures (e.g. code_upgrade). The logging levels for this logger is from the perspective of the procedure being executed. E.g. a critical log in the device logger results in the exiting of the thread and reporting the log. This means the code upgrade failed in some way. This does not impact the running of the main thread (which is what sys_logger is mean't to report on).

See the logging levels defined below from a device perspective.

debug: used for state of loops, variable contents, etc. More infor for debugging
info: normal operation details for the user
warning: odd things happening but doesn't prevent the function from failing (e.g. finding new CDP neighbours after upgrade)
error: important details that the user should know about regarding the procedure (e.g. stack member missing)
critical: forces the thread to exit because this condition would prevent the procedure from executing correctly (e.g. sync fails).